+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+- Add timing information to status nsq messages
+- Add additional status nsq messages for unpack, pack, upload
+- Allow for `npm install` retries in nsq status messages
+
 ### 2.3.1
 
 - [#33] Default `this.target` and use it as the default for `this.rootDir`.